requirements.txt
Flask
mysqlclient
Flask-SQLAlchemy
flask-marshmallow
marshmallow-sqlalchemy
flask-cors